refactor: Sections 일급 컬렉션 추가 및 추가,삭제 비지니스 로직 이동 (line -> sections)
